requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
tifffile==2021.1.14
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
